Today(1/11/25), I started my dj video game.
I had never worked with python classes before, so I did some work with that, and I am now able to make a Dj for my game, and show a blank canvas to the screen.

I have chosen to use Python for this project, not Lua!

This is the very start of this project, so there is lots more to come!